SettingsActivity.kt SettingsFragment.kt SettingsViewModel.kt MainScreen.kt TripDetailsActivity.kt Resources (strings, arrays, prefs)
- Implement new SettingsActivity and SettingsFragment using PreferenceFragmentCompat and Hilt-injected ViewModel
- Refactor unit settings to separate distance and speed units, adding support for Knots
- Rename `Units` enum to `DistanceUnits` and introduce `SpeedUnits` enum
- Add `KnotsFormatter` and update `KmhFormatter`/`MphFormatter` to display integer values
- Standardize resource naming conventions in `arrays.xml` and `strings.xml` across multiple locales (es, fr, he, iw, ru, zh)
- Update `TextIndicator` custom view to support configurable text weight via attributes
- Migrate `MainScreen` and `TripDetailsActivity` to use `PreferenceManager.getDefaultSharedPreferences`
- Update UI layouts to include bold text weights for main indicators and handle system window insets in settings
- Standardize `BaseUnitFormatter` to use green tint for unit labels
- Add `@RequiresPermission` annotations to location-dependent methods in `MainScreen`

+++ app/src/main/kotlin/presentation/units_formatters/KmhFormatter.kt
@@ -10,7 +10,7 @@ class KmhFormatter : BaseUnitFormatter() {
 
     override fun formatUnits(data: Double): SpannableString {
         // km\h = m\sec * 3.6
-        speedBuilder.replace(0, speedBuilder.length, String.format(Locale.getDefault(), "%.1f", (data * 3.6)))
+        speedBuilder.replace(0, speedBuilder.length, String.format(Locale.getDefault(), "%d", (data * 3.6).toInt()))
         return createSpannable(speedBuilder.toString())
     }
 }
