+++ app/src/main/kotlin/presentation/units_formatters/MphFormatter.kt
@@ -9,7 +9,7 @@ class MphFormatter : BaseUnitFormatter() {
         val speedBuilder = StringBuilder()
         // 1mph = 1 m/sec * 2.23694
         val mph = data * 2.23694
-        speedBuilder.replace(0, speedBuilder.length, String.format(Locale.getDefault(), "%.1f", mph))
+        speedBuilder.replace(0, speedBuilder.length, String.format(Locale.getDefault(), "%d", mph.toInt()))
         return createSpannable(speedBuilder.toString())
     }
 }
